+++ src/Stormancer.Plugins/Epic/Stormancer.Server.Plugins.Epic/Changelog.rst
@@ -16,6 +16,9 @@ Changed
 *******
 - Updated dependency to Users to support new abstraction lib.
 - Improved sandbox error logs.
+Fixed
+*****
+- Made the plugin compatible with Users 9.0.0
 
 1.1.3.2
 -------