+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/Changelog.rst
@@ -33,6 +33,7 @@ Fixed
 - When trying to start a game server, timeout if it takes more than 5s to do a docker start on an agent.
 - Fixed false positives error logs.
 - Resetting a game session force the gamesession to be completed if it wasn't already, exactly as if all players add left it. This guarantees that if only the host sent results, they would be processed on reset, triggering (for instance) a PSN Activity completion.
+- Made the plugin compatible with Users 9.0.0
 
 6.1.1.15
 ----------