+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/Changelog.rst
@@ -25,6 +25,7 @@ Fixed
 - Use PeriodicTimer in PartyAnalyticsWorker to fix an high usage CPU issue.
 - Don't create an error log if the party of a player was already destroyed when a gamesession try to update its party state.
 - Fixed issue in party size check that prevented the connecting player from being taken into account.
+- Made the plugin compatible with Users 9.0.0.
 
 Added
 *****